// src/app/shared/login/models/user.ts
export class User {
    id?: number;
    userId?: string; 
    userName?: string; 
    email?: string;
    password?: string;
    token?: string;
    role?: string;
    roleId?: number;
    head?: number; 
    userType?: string;
    changePasswordIndicator?: boolean;
    forgetPasswordIndicator?: boolean;
}